+++ templates/profile.php
@@ -10,7 +10,18 @@
 
 <div class="subMenu">
     <div class="subMenuTab"> <a href="../public/index.php?route=updatePassword">Modifier mon mot de passe</a> </div>
-    <div class="subMenuTab"> <a href="../public/index.php?route=deleteAccount">Supprimer mon compte</a> </div>
+    
+    <?php 
+    if ($this->session->get('role') === 'admin'){
+        ?>
+        <div class="subMenuTab" id="textWhite">Suppression du compte impossible</div>
+        <?php 
+    }
+    else {
+        ?>
+        <div class="subMenuTab"> <a href="../public/index.php?route=deleteAccount">Supprimer mon compte</a> </div>
+        <?php
+    } ?>  
 </div>
